Switched from basic scraping to advanced using PlayWright to simulate browser
