+++ botwatcher.py
@@ -9,8 +9,10 @@ import time
 from pathlib import Path
 
 from helpers.logging import Logger, NotificationHandler
+from helpers.browser import (
+    get_shared_bot_data
+)
 from helpers.misc import (
-    get_shared_bot_data,
     remove_prefix,
     wait_time_interval,
 )
