+++ requirements.txt
@@ -4,3 +4,4 @@ telethon==1.24.0
 beautifulsoup4==4.10.0
 aiohttp==3.8.1
 schedule==1.1.0
+#playwright==1.27.0
